feat(Cargo.toml): bump version to 0.1.11 for celestial-hub-compass package refactor(ast/mod.rs): change Operand::as_immediate return type to i32 for consistency refactor(codegen/mips/mod.rs): update Operand::as_immediate to return i32 for consistency refactor(codegen/mips/mod.rs): update load_immediate calls to accept i32 values refactor(codegen/mips/mod.rs): update load_immediate_to_new_register to accept i32 values refactor(codegen/mips/mod.rs): update Operand::LiteralU32 handling to accept i32 values

+++ src/ast/mod.rs
@@ -224,19 +224,19 @@ impl Operand {
     })
   }
 
-  pub fn as_immediate(&self) -> Result<u32, String> {
+  pub fn as_immediate(&self) -> Result<i32, String> {
     Ok(match self {
-      Operand::LiteralI8(val) => *val as u32,
-      Operand::LiteralI16(val) => *val as u32,
-      Operand::LiteralI32(val) => *val as u32,
-      Operand::LiteralI64(val) => *val as u32,
-      Operand::LiteralU8(val) => *val as u32,
-      Operand::LiteralU16(val) => *val as u32,
-      Operand::LiteralU32(val) => *val,
-      Operand::LiteralU64(val) => *val as u32,
-      Operand::LiteralF32(val) => *val as u32,
-      Operand::LiteralF64(val) => *val as u32,
-      Operand::LiteralBool(val) => *val as u32,
+      Operand::LiteralI8(val) => *val as i32,
+      Operand::LiteralI16(val) => *val as i32,
+      Operand::LiteralI32(val) => *val as i32,
+      Operand::LiteralI64(val) => *val as i32,
+      Operand::LiteralU8(val) => *val as i32,
+      Operand::LiteralU16(val) => *val as i32,
+      Operand::LiteralU32(val) => *val as i32,
+      Operand::LiteralU64(val) => *val as i32,
+      Operand::LiteralF32(val) => *val as i32,
+      Operand::LiteralF64(val) => *val as i32,
+      Operand::LiteralBool(val) => *val as i32,
       _ => return Err("Expected immediate, found register".to_string()),
     })
   }
